fix(*): structure

+++ src/main/java/gok/api/domain/dto/CreateProductRequest.java
@@ -1,6 +1,7 @@
 package gok.api.domain.dto;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
-public record CreateProductRequest(String name, Double price, String type, Optional<String> description) {
+public record CreateProductRequest(String name, BigDecimal price, String type, Optional<String> description) {
 }
