+++ src/main/java/gok/api/domain/dto/UpdateProductRequest.java
@@ -1,10 +1,11 @@
 package gok.api.domain.dto;
 
+import java.math.BigDecimal;
 import java.util.Optional;
 
 public record UpdateProductRequest(
         Optional<String> name,
-        Optional<Double> price,
+        Optional<BigDecimal> price,
         Optional<String> type,
         Optional<String> description
 ) {
